+++ src/index.css
@@ -26,7 +26,7 @@
   padding-top: 10px;
   padding-bottom: 10px;
   object-fit: contain;
-  border-bottom: 1px solid var(--color);
+  border-bottom: 1px solid rgb(219, 219, 219);
   display: flex;
   justify-content: space-between;
   align-items: center;
