+++ src/pages/Profile.jsx
@@ -6,20 +6,15 @@ import {
   Typography,
   useMediaQuery,
 } from "@mui/material";
-import { auth, db, storage } from "../lib/firebase";
-import { useEffect, useState } from "react";
+import { auth, storage } from "../lib/firebase";
+import { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
 import { FaUserCircle } from "react-icons/fa";
-import Post from "../components/Post";
-import ShareModal from "../components/ShareModal";
 import { useSnackbar } from "notistack";
+import SideBar from "../components/SideBar";
 
 function Profile() {
-  const [openShareModal, setOpenShareModal] = useState(false);
-  const [currentPostLink, setCurrentPostLink] = useState("");
-  const [postText, setPostText] = useState("");
-  const [posts, setPosts] = useState([]);
   const { name, email, avatar } = useLocation().state;
   const isNonMobile = useMediaQuery("(min-width: 768px)");
   const { enqueueSnackbar } = useSnackbar();
@@ -28,23 +23,6 @@ function Profile() {
   const [visible, setVisibile] = useState(false);
   const navigate = useNavigate();
 
-  useEffect(() => {
-    const fetchPosts = async () => {
-      const postsRef = db.collection("posts");
-      const snapshot = await postsRef.get();
-      const posts = [];
-      snapshot.forEach((doc) => {
-        if (
-          localStorage.getItem("posts") &&
-          localStorage.getItem("posts").includes(doc.id)
-        ) {
-          posts.push({ id: doc.id, post: doc.data() });
-        }
-      });
-      setPosts(posts);
-    };
-    fetchPosts();
-  }, []);
   const handleBack = () => {
     navigate("/dummygram"); // Use navigate function to change the URL
   };
@@ -90,6 +68,7 @@ function Profile() {
 
   return (
     <>
+      <SideBar />
       <Box
         width={isNonMobile ? "30%" : "70%"}
         backgroundColor="#F4EEFF"
@@ -175,38 +154,6 @@ function Profile() {
             Back
           </Button>
         </Box>
-      </Box>
-      <ShareModal
-        openShareModal={openShareModal}
-        setOpenShareModal={setOpenShareModal}
-        currentPostLink={currentPostLink}
-        postText={postText}
-      />
-      <Box>
-        <div
-          className="profile__favourites"
-          style={{ marginTop: "1.5rem", marginBottom: "1.5rem" }}
-          align="center"
-        >
-          {posts.length ? (
-            <>
-              <h1>Your Favourites</h1>
-              {posts.map(({ id, post }) => (
-                <Post
-                  key={id}
-                  postId={id}
-                  user={auth.currentUser}
-                  post={post}
-                  shareModal={setOpenShareModal}
-                  setLink={setCurrentPostLink}
-                  setPostText={setPostText}
-                />
-              ))}
-            </>
-          ) : (
-            <>You have nothing in favourites</>
-          )}
-        </div>
       </Box>
     </>
   );
